Move Condition class into conditions module.
